Split router.go (2175 lines) into 8 logical files
Pure reorganization — no code changes, just moving functions:

- router.go (252 lines): types, interface, constructor, constants
- router_dial.go (688): DialRoutes, PingRoute, fetchBestRoutes, calculateLocalRoutes
- router_serve.go (290): Serve, AcceptRoutes, serveTransportManager, serveSetup
- router_packet.go (439): all handle*Packet methods, forwardPacket
- router_gc.go (105): rulesGCLoop, rulesGC, removeRouteGroupOfRule
- router_mux_ops.go (167): appendRouteToGroup, Add/RemoveMuxRouteByTransport
- router_rules.go (214): IntroduceRules, rule CRUD, route group map operations
- router_config.go (105): setters, Close, isTpdExist
